backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "p-retrover",email = "primalretrover@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-google-genai (>=2.1.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
alembic = "^1.16.5"
psycopg2-binary = "^2.9.10"
httpx = "^0.28.1"
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"

[tool.poetry]
name = "backend"
version = "0.1.0"
readme = "README.md"
package-mode = false # Use false for non-package projects